Convert slow running valgrind heavy tests to none
The interpolator does indeed make these tests run slower and here
is why: we leverage AD computations to figure out the 'a' coefficient
contributions for some kernels. This happens irrespective of whether
we are doing residual or Jacobian evaluations. This change has a large
impact for simulations that have expensive kernels like the reynolds
stress kernel. So especially in order to make threading happy we toggle
on `ADReal::do_derivatives` during RC data evaluation so we've basically
doubled the number of Jacobian evaluations (when `line_search = 'none'`).
This would be made worse by having the default line search of bt.
Now I want to implement the capability to simultaneously evaluate
residual and Jacobian at the same time