Displaced mesh must always be an exact clone
In #24018 we checked to see whether the reference mesh base object was prepared
in `MooseMesh::prepare`, and if so then we would later also prepare the
displaced mesh. The issue with this is that even Metis may yield different
partitions for identical meshes and this plays havoc when we want to sync
vectors between undisplaced and displaced meshes. Back when we had this problem
in the past we proposed issue #12799 as a solution but ended up going with #12889
which has the same spirit as this new fix: *ensure* the displaced mesh is
an exact copy of the reference mesh. Now of course I am getting paranoid about
what can happen when we have adaptivity...

Thank you to discussion #24499 for illustrating that we had this issue
crop up again